# docs/swagger.yaml
basePath: https://api.web95.tech/api/v1
definitions:
  helpers.ErrorData:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  models.BuyRequest:
    properties:
      date:
        type: string
      name:
        type: string
      phone:
        type: string
      price:
        type: string
      username:
        type: string
    type: object
...
  models.CreateTourRequest:
    properties:
      hotels:
        items:
          $ref: '#/definitions/models.HotelRequest'
        type: array
      route_cities:
        additionalProperties:
          $ref: '#/definitions/models.TripRouteCity'
        type: object
      trip:
        $ref: '#/definitions/models.CreateTripRequest'
    type: object
  models.CreateTourResponse:
    properties:
      hotels:
        items:
          $ref: '#/definitions/models.HotelResponse'
        type: array
      routes:
        items:
          $ref: '#/definitions/models.TripRoute'
        type: array
      success:
        type: boolean
      trip:
        $ref: '#/definitions/models.Trip'
    type: object
  models.CreateTripRequest:
    properties:
      active:
        type: boolean
      booking_deadline:
        type: string
      currency:
        type: string
      departure_city:
        type: string
      description:
        type: string
      discount_percent:
        type: integer
      end_date:
        type: string
      hotels:
        items:
          $ref: '#/definitions/models.HotelAttach'
        type: array
      main:
        type: boolean
      price:
        type: number
      season:
        type: string
      start_date:
        type: string
      title:
        type: string
      trip_type:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
...
  models.HotelAttach:
    properties:
      hotel_id:
        type: integer
      nights:
        type: integer
    type: object
  models.HotelRequest:
    properties:
      city:
        type: string
      distance:
        type: number
      distance_text:
        type: string
      guests:
        type: string
      meals:
        type: string
      name:
        type: string
      stars:
        type: integer
      transfer:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
  models.HotelResponse:
    properties:
      city:
        type: string
      created_at:
        type: string
      distance:
        type: number
      distance_text:
        type: string
      guests:
        type: string
      id:
        type: integer
      meals:
        type: string
      name:
        type: string
      nights:
        type: integer
      stars:
        type: integer
      transfer:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
...
  models.Trip:
    properties:
      active:
        type: boolean
      booking_deadline:
        type: string
      buys_count:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      departure_city:
        type: string
      description:
        type: string
      discount_percent:
        type: integer
      end_date:
        type: string
      final_price:
        type: number
      hotels:
        items:
          $ref: '#/definitions/models.TripHotelWithInfo'
        type: array
      id:
        type: integer
      main:
        type: boolean
      price:
        type: number
      season:
        type: string
      start_date:
        type: string
      title:
        type: string
      trip_type:
        type: string
      updated_at:
        type: string
      urls:
        description: "\U0001F448 массив ссылок вместо photo_url"
        items:
          type: string
        type: array
      views_count:
        type: integer
    type: object
  models.TripFullResponse:
    properties:
      hotels:
        items:
          $ref: '#/definitions/models.HotelResponse'
        type: array
      routes:
        items:
          $ref: '#/definitions/models.TripRoute'
        type: array
      trip:
        $ref: '#/definitions/models.Trip'
    type: object
  models.TripFullUpdateRequest:
    properties:
      hotels:
        items:
          $ref: '#/definitions/models.TripHotel'
        type: array
      routes:
        items:
          $ref: '#/definitions/models.TripRoute'
        type: array
      trip:
        $ref: '#/definitions/models.UpdateTripRequest'
    type: object
  models.TripHotel:
    properties:
      hotel_id:
        type: integer
      nights:
        type: integer
      trip_id:
        type: integer
    type: object
  models.TripHotelWithInfo:
    properties:
      city:
        type: string
      distance:
        type: number
      hotel_id:
        type: integer
      meals:
        type: string
      name:
        type: string
      nights:
        type: integer
      rating:
        type: integer
    type: object
...
  models.TripRoute:
    properties:
      city:
        type: string
      created_at:
        type: string
      duration:
        type: string
      id:
        type: integer
      position:
        type: integer
      stop_time:
        type: string
      transport:
        type: string
      trip_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.TripRouteCity:
    properties:
      city:
        type: string
      duration:
        type: string
      stop_time:
        type: string
      transport:
        type: string
    required:
    - city
    type: object
...
  models.UpdateTripRequest:
    properties:
      active:
        type: boolean
      booking_deadline:
        type: string
      currency:
        type: string
      departure_city:
        type: string
      description:
        type: string
      discount_percent:
        type: integer
      end_date:
        type: string
      hotels:
        items:
          $ref: '#/definitions/models.HotelAttach'
        type: array
      main:
        type: boolean
      price:
        type: number
      season:
        type: string
      start_date:
        type: string
      title:
        type: string
      trip_type:
        type: string
      urls:
        description: "\U0001F448 массив ссылок"
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: API for Travel project
  title: Travel API
  version: "1.0"
paths:
...
  /admin/tours:
    post:
      consumes:
      - application/json
      description: 'Админская ручка: создаёт тур, отель и маршрут одним запросом'
      parameters:
      - description: Tour + Hotel + Route
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateTourRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateTourResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Create Tour with Hotel and Route
      tags:
      - Admin — Trips
  /admin/trips:
    post:
      consumes:
      - application/json
      description: Создание нового тура (только админ)
      parameters:
      - description: Trip data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Create trip (admin)
      tags:
      - Admin — Trips
  /admin/trips/{id}:
    delete:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Delete trip (admin)
      tags:
      - Admin — Trips
    put:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Trip update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Update trip (admin)
      tags:
      - Admin — Trips
  /admin/trips/{id}/full:
    get:
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TripFullResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Получить тур с отелями и маршрутами
      tags:
      - Admin — Trips
    put:
      consumes:
      - application/json
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Trip with hotels and routes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TripFullUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Обновить тур, отели и маршруты одной кнопкой
      tags:
      - Admin — Trips
...
  /trips:
    get:
      description: Публичный поиск туров с фильтрацией и пагинацией
      parameters:
      - description: Поиск по названию тура
        in: query
        name: title
        type: string
      - description: Город вылета
        in: query
        name: departure_city
        type: string
      - description: Тип тура
        in: query
        name: trip_type
        type: string
      - description: Сезон
        in: query
        name: season
        type: string
      - description: Город в маршруте
        in: query
        name: route_city
        type: string
      - description: Статус тура
        in: query
        name: active
        type: boolean
      - description: Дата начала с (YYYY-MM-DD)
        in: query
        name: start_after
        type: string
      - description: Дата окончания до (YYYY-MM-DD)
        in: query
        name: end_before
        type: string
      - description: Лимит (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success + items + meta
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка при получении списка туров
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: List trips
      tags:
      - Public — Trips
  /trips/{id}:
    get:
      description: Публичный просмотр тура
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Trip'
        "404":
          description: Тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при получении тура
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get trip by id
      tags:
      - Public — Trips
  /trips/{id}/buy:
    post:
      consumes:
      - application/json
      description: Отправка заявки на покупку тура в Telegram
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные покупателя
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.BuyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "404":
          description: Тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при покупке тура
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Buy trip
      tags:
      - Public — Trips
  /trips/{id}/countdown:
    get:
      description: Получить обратный отсчёт до конца бронирования
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "404":
          description: Тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Ошибка при получении тура
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get booking countdown
      tags:
      - Public — Trips
...
  /trips/main:
    get:
      description: Получить главный тур для главной страницы (только название и обратный
        отсчёт)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Главный тур не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get main trip with countdown
      tags:
      - Public — Trips
  /trips/popular:
    get:
      parameters:
      - description: Количество туров (по умолчанию 5)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Trip'
            type: array
        "500":
          description: Не удалось получить популярные туры
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Get popular trips
      tags:
      - Public — Trips
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"